refactor: Reaction 위치 수정
