+++ motimo-api/src/main/java/kr/co/api/group/docs/GroupControllerSwagger.java
@@ -11,13 +11,14 @@ import kr.co.api.group.rqrs.GroupMessageIdRs;
 import kr.co.api.group.rqrs.GroupMessageItemRs;
 import kr.co.api.group.rqrs.JoinedGroupRs;
 import kr.co.domain.common.pagination.CustomSlice;
-import kr.co.domain.reaction.ReactionType;
+import kr.co.domain.group.reaction.ReactionType;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestParam;
 
 @Tag(name = "그룹 API", description = "그룹 관련 API 목록입니다")
 public interface GroupControllerSwagger {
+
     @Operation(summary = "참여중인 그룹 목록 API", description = "참여중인 그룹을 조회합니다.")
     List<JoinedGroupRs> getJoinedGroup(UUID userId);
 
@@ -25,10 +26,12 @@ public interface GroupControllerSwagger {
     GroupIdRs joinRandomGroup(UUID userId, @RequestBody GroupJoinRq rq);
 
     @Operation(summary = "그룹 채팅 조회 API", description = "그룹 채팅을 조회합니다.")
-    CustomSlice<GroupMessageItemRs> getGroupChat(UUID groupId, @RequestParam int page, @RequestParam int size);
+    CustomSlice<GroupMessageItemRs> getGroupChat(UUID groupId, @RequestParam int page,
+            @RequestParam int size);
 
     @Operation(summary = "그룹 리액션 API", description = "그룹 리액션을 생성합니다.")
-    GroupMessageIdRs createGroupReaction(UUID userId, @PathVariable UUID messageId, @RequestParam ReactionType reactionType);
+    GroupMessageIdRs createGroupReaction(UUID userId, @PathVariable UUID messageId,
+            @RequestParam ReactionType reactionType);
 
     @Operation(summary = "그룹 멤버 조회 API", description = "그룹 멤버 목록을 조회합니다.")
     List<GroupMemberRs> getGroupMembers(UUID userId, @PathVariable UUID groupId);
